Provide an implementation for estimated memory usage.

+++ src/idt/targets/pure_cpp/hashing/public_include/reify/pure_cpp/thread_pool_cache_runner.h
@@ -148,7 +148,7 @@ class ThreadPoolCacheRunner {
     });
   }
 
-  int CacheEstimatedMemoryUsageInBytes() const {
+  int64_t CacheEstimatedMemoryUsageInBytes() const {
     return cache_.EstimatedMemoryUsageInBytes();
   }
